Move enums out into their own file.
