+++ src/main/groovy/com/github/alexfu/Version.groovy
@@ -2,9 +2,6 @@ package com.github.alexfu
 
 import groovy.json.JsonOutput
 import groovy.json.JsonSlurper
-import groovy.json.internal.LazyMap
-
-import static com.github.alexfu.AndroidAutoVersionPlugin.VersionType.*
 
 class Version {
     int buildNumber = 0
@@ -30,8 +27,8 @@ class Version {
         return formatter.call(major, minor, patch, buildNumber)
     }
 
-    String versionNameForFlavor(AndroidAutoVersionPlugin.VersionFlavor flavor) {
-        if (flavor == AndroidAutoVersionPlugin.VersionFlavor.BETA) {
+    String versionNameForFlavor(VersionFlavor flavor) {
+        if (flavor == VersionFlavor.BETA) {
             return betaVersionName()
         }
         return releaseVersionName()
@@ -57,22 +54,22 @@ class Version {
                                   buildNumber: buildNumber])
     }
 
-    void update(AndroidAutoVersionPlugin.VersionType type) {
+    void update(VersionType type) {
         buildNumber += 1
         revision += 1
         switch (type) {
-            case MAJOR:
+            case VersionType.MAJOR:
                 revision = 0
                 patch = 0
                 minor = 0
                 major += 1
                 break
-            case MINOR:
+            case VersionType.MINOR:
                 revision = 0
                 patch = 0
                 minor += 1
                 break
-            case PATCH:
+            case VersionType.PATCH:
                 revision = 0
                 patch += 1
                 break
