+++ src/main/groovy/com/github/alexfu/VersionFlavor.groovy
@@ -1,0 +1,11 @@
+package com.github.alexfu
+
+enum VersionFlavor {
+    RELEASE("Release"), BETA("Beta")
+
+    private final String name
+
+    private VersionFlavor(String name) {
+        this.name = name
+    }
+}
